+++ TestTaskCalculatorWPF/ICalculator.cs
@@ -2,9 +2,9 @@
 {
     public interface ICalculator
     {
-        public void Plus(double number);
-        public void Minus(double number);
-        public void Div(double number);
-        public void Multiply(double number);
+        public void Plus(decimal number);
+        public void Minus(decimal number);
+        public void Div(decimal number);
+        public void Multiply(decimal number);
     }
 }
